Create RaspiDisplay.py
This code will continuously read the values from the sensors and print them to the console. You'll need to define the read_humidity_sensor, read_pH_sensor, read_light_sensor, and read_nutrient_sensor functions to actually read the values from the sensors. The specific details of how to do this will depend on the sensors you are using.